refactor unit tests move use storage object from film storage to service

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,8 +2,10 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.Collection;
 import java.util.List;
@@ -12,6 +14,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class FilmService {
     private final FilmStorage storage;
+    private final UserStorage userStorage;
+
 
     public Collection<Film> getFilms() {
         return storage.getFilms();
@@ -26,10 +30,16 @@ public class FilmService {
     }
 
     public void addLike(Long filmId, Long userId) {
+        if (userStorage.findUserById(userId).isEmpty()) {
+            throw new NotFoundException("the is no user with id: " + userId);
+        }
         storage.addLike(filmId, userId);
     }
 
     public void deleteLike(Long filmId, Long userId) {
+        if (userStorage.findUserById(userId).isEmpty()) {
+            throw new NotFoundException("the is no user with id: " + userId);
+        }
         storage.deleteLike(filmId, userId);
     }
